comment for config files (#63)
